+++ app/services/delete_sheet.rb
@@ -9,10 +9,27 @@ module Vitae
       end
     end
 
-    def self.call(auth:, file_id:)
-      # raise ForbiddenError unless auth[:scope].can_write?('sheets')
+    class NotFoundError < StandardError
+      def message
+        'Sheet not found'
+      end
+    end
 
-      Sheet.where(file_id: file_id).destroy
+    def self.call(auth:, file_id:)
+      raise ForbiddenError unless auth
+
+      sheet = Sheet.first(file_id: file_id)
+      raise NotFoundError unless sheet
+
+      policy = SheetPolicy.new(
+        account: auth[:account],
+        sheet: sheet,
+        auth_scope: auth[:scope]
+       )
+
+      raise ForbiddenError unless policy.can_delete?
+
+      sheet.destroy
       GoogleSheets.new.delete_sheet(file_id: file_id)
     end
   end
